Revert "feat: added LearnerPathways in search/all endpoint (#3277)"
This reverts commit 6b420920c90a7fbfd991fff0173c701d73781a68.
