+++ course_discovery/apps/course_metadata/search_indexes/documents/__init__.py
@@ -1,7 +1,6 @@
 from .course import CourseDocument
 from .course_run import CourseRunDocument
-from .learner_pathway import LearnerPathwayDocument
 from .person import PersonDocument
 from .program import ProgramDocument
 
-__all__ = ('CourseRunDocument', 'CourseDocument', 'LearnerPathwayDocument', 'PersonDocument', 'ProgramDocument')
+__all__ = ('CourseRunDocument', 'CourseDocument', 'PersonDocument', 'ProgramDocument')
