+++ course_discovery/apps/learner_pathway/tests/factories.py
@@ -1,7 +1,6 @@
 import factory
 from factory.fuzzy import FuzzyText
 
-from course_discovery.apps.core.tests.factories import PartnerFactory
 from course_discovery.apps.course_metadata.tests.factories import CourseFactory, ProgramFactory
 from course_discovery.apps.learner_pathway.models import (
     LearnerPathway, LearnerPathwayCourse, LearnerPathwayProgram, LearnerPathwayStep
@@ -13,7 +12,6 @@ class LearnerPathwayFactory(factory.django.DjangoModelFactory):
         model = LearnerPathway
 
     name = FuzzyText(prefix='learner-pathway-name-')
-    partner = factory.SubFactory(PartnerFactory)
     uuid = factory.Faker('uuid4')
     banner_image = factory.django.ImageField()
     overview = FuzzyText(prefix='learner-pathway-overview-')
